Implement go to selections link and move to common (#840)
